% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\alias{clean.character}
\alias{clean.factor}
\alias{clean.data.frame}
\title{Clean punctuated headers and text variables}
\usage{
clean(x, keep = NULL, add_repl = NULL, ...)

\method{clean}{character}(x, keep = NULL, add_repl = NULL)

\method{clean}{factor}(x, keep = NULL, add_repl = NULL,
  byref = FALSE)

\method{clean}{data.frame}(x, cols = TRUE, vars = FALSE, keep = NULL,
  add_repl = NULL, byref = FALSE)
}
\arguments{
\item{x}{chr, factor or data.frame.}

\item{keep}{vector of characters to preserve}

\item{add_repl}{named vector of additional replacements desired. Overrules \code{keep} and default replacements}

\item{...}{additional parameters to factor and data.frame S3 methods. See details.}
}
\description{
Replace or drop UTF-8 characters and punctuation in headers or/and in char/factor variables.
}
\details{
For data.frame's, \code{byref = TRUE} allows to replace characters by reference. 

Additionally, booleans \code{cols} and \code{vars} indicate if replacements should take place in columns and variables, respectively. By default, \code{cols = TRUE} 
and \code{vars = FALSE}.
}
