% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{insp_outlier}
\alias{insp_outlier}
\title{Time series outlier inspection}
\usage{
insp_outlier(x, ...)
}
\arguments{
\item{x}{ts or data.frame}

\item{window_size}{integer window size. Determines how many observations near an outlier are used to calculate the input value.}

\item{anom_method}{"gesd" or "iqr", argument passed to anomalize::anomalize().}

\item{\code{...}}{Additional arguments for the data.frame case. See details.}
}
\description{
Time series outlier inspection
}
\details{
When \code{margin = 1}, set \code{idcol} to an integer pointing out the index column that identifies the time series.
When \code{margin = 2}, set \code{dtcol} to an integer pointing out the date column. 
.parallel TRUE/FALSE if tasks should be parallelized. If \code{TRUE}, uses \code{future} and \code{furrr} packages.
out_format "wide" or "long", indicating output format. If \code{margin = 2} and there are many time series, \code{out_format = "long"} is easier to handle than wide format.
}
