% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazonalidade.R
\name{inspeciona_sazonalidade}
\alias{inspeciona_sazonalidade}
\title{Inspecao de sazonalidade para series temporais}
\usage{
inspeciona_sazonalidade(x, ...)
}
\arguments{
\item{x}{Um objeto ts ou data.frame}

\item{\code{tendencia}}{\code{TRUE/FALSE} indicando se o metodo deve ser aplicado considerando series com tendencia.}

\item{\code{sentido}}{Quando x eh data.frame, \code{sentido = 1} quando as series temporais estao por linha; e \code{sentido = 2} caso estejam por colunas.}

\item{\code{periodo}}{\code{'mes','semana','dia','ano','trimestre','quadrimestre','semestre','qualquer'}.}

\item{\code{...}}{Outros argumentos para quando \code{x} for um data.frame (\code{paralelo, idcol, dtcol}). Veja a seccao de detalhes.}
}
\description{
Inspecao de sazonalidade para series temporais
}
\details{
Quando \code{sentido = 1}, \code{idcol} especifica o indice da coluna que possui os identificadores das series. 
Quando \code{sentido = 2}, \code{dtcol} especifica o indice da coluna de datas.
}
