% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na_input}
\alias{na_input}
\title{NA imputation}
\usage{
na_input(x, how = "median", window = Inf, ...)
}
\arguments{
\item{x}{Vector to input data.}

\item{how}{Inputation type. One of 'mean', 'median', 
'locf' (Last Observation Carried Forward), 
'nocb' (Next Observation Carried Backward), 
'lin_interp' (linear interpolation) or 
'cub_spline' (cubic spline)}

\item{window}{Window size. Used when \code{how} equals 'mean' or 'median'.}

\item{...}{Further arguments to zoo functions na.locf, na.approx or na.spline.}
}
\description{
Wraps and extends \code{zoo} package's NA imputation functions.
}
