% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{subs_any}
\alias{subs_any}
\title{Fast value replacement}
\usage{
subs_any(x, byref = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{byref}{\code{TRUE/FALSE} indicating if data imputation should be made by reference.}

\item{\code{...}}{Syntax for replacements. See details.}
}
\description{
Replaces specified values for any other in data.frame variables. Changes are made given column names 
or regular expressions along with its replacements in a single call.
}
\details{
Replacements must be as follows: \code{subs_any(x, byref, var_1 = list("?",NA), var_2 = list("#", 9999))}.
}
