% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{subs_na}
\alias{subs_na}
\title{Substituiccao rapida de NA}
\usage{
subs_na(x, byref = FALSE, ...)
}
\arguments{
\item{x}{Um data.frame}

\item{byref}{\code{TRUE/FALSE} indicando se a imputacao de dados deve ser feita por referencia ou retornando uma nova tabela.}

\item{\code{...}}{Substituicoes a serem feitas. Veja a seccao de detalhes.}
}
\description{
Substitui NA por valores de interesse. Eh possivel especificar nomes de colunas e/ou expressoes regulares
com seus respectivos novos valores numa so chamada.
}
\details{
As substituicoes devem ser da forma: \code{subs_na(x, byref, nome_1 = "?", nome_2 = 9999, "[0-9]{2}$" = 0)} ou
encapsuladas por \code{list()}, como: \code{subs_na(x, byref, list(nome_1 = "?", nome_2 = 9999, "[0-9]{2}$" = 0))}.
}
