% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{subs_utf8_punct}
\alias{subs_utf8_punct}
\title{Substitui caracteres com codifica\u00e7\u00e3o UTF-8}
\usage{
subs_utf8_punct(x, manter = NULL, ...)
}
\arguments{
\item{x}{chr, factor ou data.frame.}

\item{manter}{vetor com caracteres que devem ser mantidos.}

\item{...}{par\u00e2metros adicionais para classes factor e data.frame, veja Detalhes.}
}
\description{
Substitui caracteres com codifica\u00e7\u00e3o UTF-8
}
\details{
Quando trabalhando com factor ou data.frame, \u00e9 poss\u00edvel fazer a substitui\u00e7\u00e3o por
refer\u00eancia com \code{byref = TRUE}. Para data.frame, os booleanos \code{cols} e \code{vars} indicam se a
substitui\u00e7\u00e3o deve ser feitas nos nomes das vari\u00e1veis e nos valores delas, respectivamente.
}
