% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{insp_outlier}
\alias{insp_outlier}
\title{Inspeccao de outliers para series temporais}
\usage{
insp_outlier(x, ...)
}
\arguments{
\item{x}{Um objeto ts ou data.frame}

\item{janela}{Inteiro indicando tamanho da janela movel para calculo dos inputs de outliers.}

\item{\code{anom_method}}{chr indicando "gesd" ou "iqr", argumento da funcao anomalize::anomalize().}

\item{\code{...}}{Parametros adicionais para quando \code{x} eh um data.frame. Veja detalhes.}
}
\description{
Inspeccao de outliers para series temporais
}
\details{
Quando \code{sentido = 1}, \code{idcol} especifica o indice da coluna que possui os identificadores das series. 
Quando \code{sentido = 2}, \code{dtcol} especifica o indice da coluna de datas.
\code{paralelo} TRUE/FALSE indicando se o processo deve ser feito em paralelo. Quando \code{TRUE}, eh feito por meio dos pacotes \code{future} e \code{furrr}.
\code{out_format} chr "wide" ou "long", indicando formato da tabela de saida. Se \code{sentido = 2} e a dimensao das series for alta, indica-se \code{out_format = "long"}.
}
