% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{insp_seasonality}
\alias{insp_seasonality}
\title{Time series seasonality inspection}
\usage{
insp_seasonality(x, ...)
}
\arguments{
\item{x}{ts or data.frame}

\item{trend}{\code{TRUE/FALSE} indicating if input time series has trend.}

\item{margin}{For data.frames, \code{margin = 1} if multiple time series are framed row-wise; while \code{margin = 2} if column-wise.}

\item{periodicity}{\code{'month','week','day','year','three_months','four_months','six_months','any'}.}

\item{\code{...}}{Additional arguments for the data.frame case. See details.}
}
\description{
Time series seasonality inspection
}
\details{
\itemize{
 \item When \code{margin = 1}, set \code{idcol} to an integer pointing out the index column that identifies the time series;
 \item When \code{margin = 2}, set \code{dtcol} to an integer pointing out the date column;
 \item .parallel TRUE/FALSE if tasks should be parallelized. If \code{TRUE}, uses \code{future} and \code{furrr} packages.
}
}
