% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sazonalidade.R
\name{inspeciona_sazonalidade}
\alias{inspeciona_sazonalidade}
\title{Inspeção de sazonalidade para séries temporais}
\usage{
inspeciona_sazonalidade(x, ...)
}
\description{
Inspeção de sazonalidade para séries temporais
}
