% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na_prop}
\alias{na_prop}
\title{Descreve proporção de NA por variáveis de uma tabela}
\usage{
na_prop(dt, drop = FALSE, corte = if (drop) 0.5 else 0)
}
\arguments{
\item{dt}{Um data.table}

\item{drop}{TRUE/FALSE para remover variável que possui mais do que proporção \code{corte} de NA}

\item{corte}{Numérico entre 0 e 1. Com \code{drop = FALSE}, só trunca a descrição; com \code{drop = TRUE}, é a condição para remover variáveis.}
}
\value{
Tabela com descrição de NA ou data.frame original sem variáveis excluídas.
}
\description{
Descreve proporção de NA por variáveis de uma tabela
}
