% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na_prop}
\alias{na_prop}
\title{Column-wise NA summary}
\usage{
na_prop(x, min_prop = 0)
}
\arguments{
\item{x}{A data.frame}

\item{min_prop}{Numeric between 0 and 1 to limit table output}
}
\value{
Table with column-wise NA summary
}
\description{
NA summary of data.frame variables.
}
