% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.R
\name{na_prop}
\alias{na_prop}
\title{NA por colunas}
\usage{
na_prop(dt, corte = 0)
}
\arguments{
\item{corte}{Numerico entre 0 e 1 para limitar \code{print} da tabela com descricao.}

\item{x}{Uma tabela}
}
\value{
Tabela com descrição de NAs por coluna.
}
\description{
Descreve proporcao de NA por variáveis de um data.frame
}
