% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{subs_any}
\alias{subs_any}
\title{Substituiccao rapida de valores}
\usage{
subs_any(x, byref = FALSE, ...)
}
\arguments{
\item{x}{Um data.frame}

\item{byref}{\code{TRUE/FALSE} indicando se a imputacao de dados deve ser feita por referencia ou retornando uma nova tabela.}

\item{\code{...}}{Substituicoes a serem feitas. Veja a seccao de detalhes.}
}
\description{
Substitui informacoes por valores de interesse. Eh possivel especificar nomes de colunas e/ou expressoes regulares
com seus respectivos novos valores numa so chamada.
}
\details{
As substituicoes devem ser da forma: \code{subs_any(x, byref, nome_1 = list("?",NA), nome_2 = list("#", 9999))}.
}
