% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subs.R
\name{subs_na}
\alias{subs_na}
\title{Fast NA replacement}
\usage{
subs_na(x, byref = FALSE, ...)
}
\arguments{
\item{x}{A data.frame}

\item{byref}{\code{TRUE/FALSE} indicating if data imputation should be made by reference.}

\item{\code{...}}{Syntax for replacements. See details.}
}
\description{
Replaces NA for other values in data.frame variables. Changes are made given column names or regular
expressions along with its replacements in a single call.
}
\details{
Replacements must be either \code{subs_na(x, byref, var_1 = "?", var_2 = 9999, "[0-9]{2}$" = 0)} or
wrapped by \code{list()}: \code{subs_na(x, byref, list(var_1 = "?", var_2 = 9999, "[0-9]{2}$" = 0))}.
}
